+++ CppSource.pro
@@ -1,0 +1,10 @@
+TEMPLATE = app
+CONFIG += console c++11
+CONFIG -= app_bundle
+CONFIG -= qt
+
+SOURCES += \
+        main.cpp
+
+HEADERS += \
+    all.h
