+++ all.h
@@ -73,5 +73,5 @@
 #include <condition_variable>
 #include <future>
 
-using namespace std::literals;
+// using namespace std::literals;
 using namespace std;
